fix: rename hook files to the appropriate hook name & updated the imports
