+++ src/hooks/usePlatformHardwareLogs/usePlatformHardwareLogs.ts
